# Fix `Quad.Contains()` failing for zero-sized drawables