--- a/osu.Framework.Tests/Primitives/QuadTest.cs
+++ b/osu.Framework.Tests/Primitives/QuadTest.cs
@@ -7,6 +7,7 @@ using System.Collections;
 using NUnit.Framework;
 using osu.Framework.Extensions.MatrixExtensions;
 using osu.Framework.Graphics.Primitives;
+using osu.Framework.Utils;
 using osuTK;
 
 namespace osu.Framework.Tests.Primitives
@@ -30,6 +31,13 @@ namespace osu.Framework.Tests.Primitives
                 new Vector2(5, 1),
                 new Vector2(0, 5),
                 new Vector2(7, 7)
+            ),
+            // Very small quad
+            new Quad(
+                new Vector2(-Precision.FLOAT_EPSILON, -Precision.FLOAT_EPSILON),
+                new Vector2(Precision.FLOAT_EPSILON, -Precision.FLOAT_EPSILON),
+                new Vector2(-Precision.FLOAT_EPSILON, Precision.FLOAT_EPSILON),
+                new Vector2(Precision.FLOAT_EPSILON, Precision.FLOAT_EPSILON)
             )
         };
 
@@ -69,6 +77,49 @@ namespace osu.Framework.Tests.Primitives
             Assert.That(quad.Contains((quad.BottomLeft + quad.BottomRight) / 2), Is.True);
             Assert.That(quad.Contains((quad.TopLeft + quad.BottomLeft) / 2), Is.True);
             Assert.That(quad.Contains((quad.TopRight + quad.BottomRight) / 2), Is.True);
+        }
+
+        [Test]
+        public void TestContains_ZeroHeightQuad()
+        {
+            var quad = new Quad(-2, 0, 4, 0);
+
+            Assert.That(quad.Contains(new Vector2(-5, -5)), Is.False);
+            Assert.That(quad.Contains(new Vector2(-5, 5)), Is.False);
+            Assert.That(quad.Contains(new Vector2(5, -5)), Is.False);
+            Assert.That(quad.Contains(new Vector2(5, 5)), Is.False);
+
+            Assert.That(quad.Contains(new Vector2(-2, 0)), Is.True);
+            Assert.That(quad.Contains(new Vector2(0, 0)), Is.True);
+            Assert.That(quad.Contains(new Vector2(2, 0)), Is.True);
+        }
+
+        [Test]
+        public void TestContains_ZeroWidthQuad()
+        {
+            var quad = new Quad(0, -2, 0, 4);
+
+            Assert.That(quad.Contains(new Vector2(-5, -5)), Is.False);
+            Assert.That(quad.Contains(new Vector2(-5, 5)), Is.False);
+            Assert.That(quad.Contains(new Vector2(5, -5)), Is.False);
+            Assert.That(quad.Contains(new Vector2(5, 5)), Is.False);
+
+            Assert.That(quad.Contains(new Vector2(0, -2)), Is.True);
+            Assert.That(quad.Contains(new Vector2(0, 0)), Is.True);
+            Assert.That(quad.Contains(new Vector2(0, 2)), Is.True);
+        }
+
+        [Test]
+        public void TestContains_ZeroSizedQuad()
+        {
+            var quad = new Quad(0, 0, 0, 0);
+
+            Assert.That(quad.Contains(new Vector2(-5, -5)), Is.False);
+            Assert.That(quad.Contains(new Vector2(-5, 5)), Is.False);
+            Assert.That(quad.Contains(new Vector2(5, -5)), Is.False);
+            Assert.That(quad.Contains(new Vector2(5, 5)), Is.False);
+
+            Assert.That(quad.Contains(new Vector2(0, 0)), Is.True);
         }
 
         private class AreaTestData