As reported [on discord](https://discord.com/channels/188630481301012481/188630652340404224/1085495571202183208). Regressed in #5670.

Test coverage added to cover this case and a few others that did not end up needing targeted fixes; the fix pretty much targets the zero-sized case directly.

# Benchmarking results

## Before

|   Method | NumPoints |        Mean |      Error |     StdDev | Allocated |
|--------- |---------- |------------:|-----------:|-----------:|----------:|
| Contains |        10 |    64.12 ns |   1.296 ns |   2.056 ns |         - |
| Contains |       100 |   580.51 ns |  11.631 ns |  20.674 ns |         - |
| Contains |      1000 | 5,989.39 ns | 118.540 ns | 188.017 ns |         - |

## After

|   Method | NumPoints |        Mean |     Error |     StdDev | Allocated |
|--------- |---------- |------------:|----------:|-----------:|----------:|
| Contains |        10 |    73.32 ns |  1.353 ns |   1.661 ns |         - |
| Contains |       100 |   673.57 ns |  9.624 ns |   7.514 ns |         - |
| Contains |      1000 | 6,975.66 ns | 92.932 ns | 114.129 ns |         - |

tl;dr: About a 16% performance cost, although this implementation is still [better than the previous](https://github.com/ppy/osu-framework/pull/5670#issue-1614424916).